server/pyproject.toml
[project]
name = "fastapi-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alibabacloud-bailian20231229==1.8.2",
    "bs4>=0.0.2",
    "dashscope>=1.23.5",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.0.5",
    "langchain-experimental>=0.3.4",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "langgraph-checkpoint-redis>=0.0.7",
    "openai>=1.78.0",
    "psutil>=7.0.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.11.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "selenium>=4.34.2",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

